refactor: extract subscribe func
